feat: choose side of data displayed per line + write!
This commit adds rudimentary support for chosing
the size of chunk of data displayed per line.
For example, one can choose a u8 or u16 chunk of
data to be displayed per line.

The commit also replaces `print!` family of calls with
`write!()` calls so that if used with a pager, it doesn't
panic with `pipe closed` error when exiting the pager.